requirements.txt
python-dotenv==0.21.
python-telegram-bot==13.2.0
redis==5.0.1